# backend/src/docs/api.yaml
openapi: 3.0.0
info:
  title: TodoApp API
  version: 1.0.0
  description: |
    # TodoApp API Documentation

    A comprehensive REST API for managing todos with user authentication, built with Node.js, Express, TypeScript, and MySQL.

    ## Features
    - 🔐 **JWT Authentication** - Secure user authentication with access and refresh tokens
    - 🌐 **Google OAuth** - Sign in with Google integration
    - 📝 **Todo Management** - Full CRUD operations for todos
    - 🔒 **Password Reset** - Secure password reset via email
    - 📊 **Filtering** - Filter todos by status (all, completed, upcoming)
    - 🛡️ **Security** - Rate limiting, CORS, helmet, input sanitization
    - 📈 **Monitoring** - Prometheus metrics endpoint

    ## Authentication
    Most endpoints require authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    API is rate limited to 100 requests per 15 minutes per IP address.

    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "error": "Error message",
      "details": "Additional error details (optional)"
    }
    ```
  contact:
    name: TodoApp Support
    email: support@todoapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique todo identifier
          example: "123e4567-e89b-12d3-a456-426614174001"
        name:
          type: string
          maxLength: 100
          description: Todo title/name
          example: "Complete project documentation"
        shortDescription:
          type: string
          maxLength: 500
          description: Detailed description of the todo
          example: "Write comprehensive API documentation including all endpoints, examples, and error codes"
        dateTime:
          type: string
          format: date-time
          description: Due date and time for the todo
          example: "2024-01-20T15:30:00Z"
        isDone:
          type: boolean
          description: Whether the todo is completed
          example: false
          default: false
        userId:
          type: string
          format: uuid
          description: ID of the user who owns this todo
          example: "123e4567-e89b-12d3-a456-426614174000"
        createdAt:
          type: string
          format: date-time
          description: Todo creation timestamp
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Todo last update timestamp
          example: "2024-01-15T10:30:00Z"
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Todo soft deletion timestamp (null if not deleted)
          example: null
      required: [id, name, shortDescription, dateTime, isDone, userId]
...
    CreateTodoRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Todo title/name
          example: "Complete project documentation"
        shortDescription:
          type: string
          maxLength: 500
          description: Detailed description of the todo
          example: "Write comprehensive API documentation including all endpoints, examples, and error codes"
        dateTime:
          type: string
          format: date-time
          description: Due date and time for the todo
          example: "2024-01-20T15:30:00Z"
      required: [name, shortDescription, dateTime]
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid credentials"
        details:
          type: string
          description: Additional error details
          example: "The provided email or password is incorrect"
      required: [error]
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              property:
                type: string
                example: "email"
              constraints:
                type: object
                example:
                  isEmail: "email must be an email"
          description: Array of validation errors
      required: [errors]
...
paths:
...
  /api/todos:
    post:
      summary: Create a new todo
      description: |
        Creates a new todo item for the authenticated user. All input is sanitized to prevent XSS attacks.
        
        **Features:**
        - Input sanitization with sanitize-html
        - Automatic user association
        - Validation of required fields
        - Date/time scheduling support
        
        **Authentication Required:**
        This endpoint requires a valid JWT access token in the Authorization header.
      tags: [Todos]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
            examples:
              simple_todo:
                summary: Simple todo example
                value:
                  name: "Complete project documentation"
                  shortDescription: "Write comprehensive API documentation including all endpoints, examples, and error codes"
                  dateTime: "2024-01-20T15:30:00Z"
              urgent_todo:
                summary: Urgent todo with near deadline
                value:
                  name: "Fix critical bug"
                  shortDescription: "Resolve the authentication issue causing login failures"
                  dateTime: "2024-01-16T09:00:00Z"
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
              example:
                id: "123e4567-e89b-12d3-a456-426614174001"
                name: "Complete project documentation"
                shortDescription: "Write comprehensive API documentation including all endpoints, examples, and error codes"
                dateTime: "2024-01-20T15:30:00Z"
                isDone: false
                userId: "123e4567-e89b-12d3-a456-426614174000"
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-15T10:30:00Z"
                deletedAt: null
        '400':
          description: Validation error - missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                errors:
                  - property: "name"
                    constraints:
                      isNotEmpty: "name should not be empty"
                  - property: "dateTime"
                    constraints:
                      isDate: "dateTime must be a valid date"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Authorization header required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to create todo"

    get:
      summary: Get user's todos with optional filtering
      description: |
        Retrieves all todos for the authenticated user with optional status filtering. Results are ordered by due date (ascending).
        
        **Filtering Options:**
        - No filter: Returns all todos (completed and upcoming)
        - `status=COMPLETED`: Returns only completed todos (isDone = true)
        - `status=UPCOMING`: Returns only upcoming/pending todos (isDone = false)
        
        **Features:**
        - User-specific todos only (data isolation)
        - Ordered by due date (earliest first)
        - Soft-deleted todos are excluded
        - Optional status filtering
        
        **Authentication Required:**
        This endpoint requires a valid JWT access token in the Authorization header.
      tags: [Todos]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [COMPLETED, UPCOMING]
          required: false
          description: Filter todos by completion status
          examples:
            all:
              summary: Get all todos
              value: ""
            completed:
              summary: Get completed todos only
              value: "COMPLETED"
            upcoming:
              summary: Get upcoming todos only
              value: "UPCOMING"
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
              examples:
                all_todos:
                  summary: All todos (mixed status)
                  value:
                    - id: "123e4567-e89b-12d3-a456-426614174001"
                      name: "Complete project documentation"
                      shortDescription: "Write comprehensive API documentation"
                      dateTime: "2024-01-20T15:30:00Z"
                      isDone: false
                      userId: "123e4567-e89b-12d3-a456-426614174000"
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-15T10:30:00Z"
                      deletedAt: null
                    - id: "123e4567-e89b-12d3-a456-426614174002"
                      name: "Review code changes"
                      shortDescription: "Review and approve pending pull requests"
                      dateTime: "2024-01-18T14:00:00Z"
                      isDone: true
                      userId: "123e4567-e89b-12d3-a456-426614174000"
                      createdAt: "2024-01-14T09:15:00Z"
                      updatedAt: "2024-01-17T16:45:00Z"
                      deletedAt: null
                completed_only:
                  summary: Completed todos only
                  value:
                    - id: "123e4567-e89b-12d3-a456-426614174002"
                      name: "Review code changes"
                      shortDescription: "Review and approve pending pull requests"
                      dateTime: "2024-01-18T14:00:00Z"
                      isDone: true
                      userId: "123e4567-e89b-12d3-a456-426614174000"
                      createdAt: "2024-01-14T09:15:00Z"
                      updatedAt: "2024-01-17T16:45:00Z"
                      deletedAt: null
                empty_list:
                  summary: No todos found
                  value: []
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Authorization header required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to retrieve todos"
  /api/todos/{id}:
    put:
      summary: Update an existing todo
      description: |
        Updates an existing todo item. Only the todo owner can update their todos. All fields are optional - only provided fields will be updated.
        
        **Updatable Fields:**
        - `name`: Todo title/name (sanitized)
        - `shortDescription`: Detailed description (sanitized)
        - `dateTime`: Due date and time
        - `isDone`: Completion status (boolean)
        
        **Features:**
        - Partial updates (only send fields you want to change)
        - Input sanitization for text fields
        - User ownership validation
        - UUID validation for todo ID
        
        **Authentication Required:**
        This endpoint requires a valid JWT access token in the Authorization header.
      tags: [Todos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo to update
          example: "123e4567-e89b-12d3-a456-426614174001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
            examples:
              complete_todo:
                summary: Mark todo as completed
                value:
                  isDone: true
              update_details:
                summary: Update todo details
                value:
                  name: "Complete project documentation (Updated)"
                  shortDescription: "Write comprehensive API documentation with examples and testing guide"
                  dateTime: "2024-01-21T15:30:00Z"
              reschedule:
                summary: Reschedule todo
                value:
                  dateTime: "2024-01-25T10:00:00Z"
              full_update:
                summary: Update all fields
                value:
                  name: "Finalize project documentation"
                  shortDescription: "Complete and review all API documentation before release"
                  dateTime: "2024-01-22T16:00:00Z"
                  isDone: false
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
              example:
                id: "123e4567-e89b-12d3-a456-426614174001"
                name: "Complete project documentation (Updated)"
                shortDescription: "Write comprehensive API documentation with examples and testing guide"
                dateTime: "2024-01-21T15:30:00Z"
                isDone: false
                userId: "123e4567-e89b-12d3-a456-426614174000"
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-16T14:20:00Z"
                deletedAt: null
        '400':
          description: Invalid todo ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid ID"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Authorization header required"
        '404':
          description: Todo not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Todo not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to update todo"

    delete:
      summary: Delete a todo (soft delete)
      description: |
        Soft deletes a todo item. The todo is not permanently removed from the database but marked as deleted and will not appear in future queries.
        
        **Soft Delete Benefits:**
        - Data recovery possible if needed
        - Audit trail preservation
        - Referential integrity maintained
        - Compliance with data retention policies
        
        **Features:**
        - User ownership validation
        - UUID validation for todo ID
        - Audit logging of deletion
        - Immediate removal from user's todo list
        
        **Authentication Required:**
        This endpoint requires a valid JWT access token in the Authorization header.
      tags: [Todos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo to delete
          example: "123e4567-e89b-12d3-a456-426614174001"
      responses:
        '204':
          description: Todo deleted successfully (no content returned)
        '400':
          description: Invalid todo ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid ID"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Authorization header required"
        '404':
          description: Todo not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Todo not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to delete todo" 